Move tileCover to its own file
